refactor: extract files and use exceptions
