+++ .gitignore
@@ -1,1 +1,2 @@
 .vscode
+src/__pycache__
